feat!: move HonkProof and the corresponding field serde to noir-types
