+++ co-noir/co-builder/src/serialize.rs
@@ -1,11 +1,7 @@
-use crate::{HonkProofError, HonkProofResult, TranscriptFieldType, prelude::HonkCurve};
+use crate::{TranscriptFieldType, prelude::HonkCurve};
 use ark_ec::{AffineRepr, CurveConfig, CurveGroup};
 use ark_ff::{Field, PrimeField};
-use num_bigint::BigUint;
-
-pub struct Serialize<F: Field> {
-    phantom: std::marker::PhantomData<F>,
-}
+use noir_types::SerializeF;
 
 pub struct SerializeC<C: CurveGroup> {
     phantom: std::marker::PhantomData<C>,
@@ -13,131 +9,6 @@ pub struct SerializeC<C: CurveGroup> {
 
 pub struct SerializeP<P: CurveGroup> {
     phantom: std::marker::PhantomData<P>,
-}
-
-impl<F: Field> Serialize<F> {
-    const NUM_64_LIMBS: u32 = <F::BasePrimeField as PrimeField>::MODULUS_BIT_SIZE.div_ceil(64);
-    const FIELDSIZE_BYTES: u32 = Self::NUM_64_LIMBS * 8;
-    const VEC_LEN_BYTES: u32 = 4;
-
-    // TODO maybe change to impl Read?
-    pub fn from_buffer(buf: &[u8], size_included: bool) -> HonkProofResult<Vec<F>> {
-        let size = buf.len();
-        let mut offset = 0;
-
-        // Check sizes
-        let num_elements = if size_included {
-            let num_elements =
-                (size - Self::VEC_LEN_BYTES as usize) / Self::FIELDSIZE_BYTES as usize;
-            if num_elements * Self::FIELDSIZE_BYTES as usize + Self::VEC_LEN_BYTES as usize != size
-            {
-                return Err(HonkProofError::InvalidProofLength);
-            }
-
-            let read_num_elements = Self::read_u32(buf, &mut offset);
-            if read_num_elements != num_elements as u32 {
-                return Err(HonkProofError::InvalidProofLength);
-            }
-            num_elements
-        } else {
-            let num_elements = size / Self::FIELDSIZE_BYTES as usize;
-            if num_elements * Self::FIELDSIZE_BYTES as usize != size {
-                return Err(HonkProofError::InvalidProofLength);
-            }
-            num_elements
-        };
-
-        // Read data
-        let mut res = Vec::with_capacity(num_elements);
-        for _ in 0..num_elements {
-            res.push(Self::read_field_element(buf, &mut offset));
-        }
-        debug_assert_eq!(offset, size);
-        Ok(res)
-    }
-
-    pub(crate) fn field_size() -> usize {
-        Self::FIELDSIZE_BYTES as usize * F::extension_degree() as usize
-    }
-
-    pub fn to_buffer(buf: &[F], include_size: bool) -> Vec<u8> {
-        let total_size = buf.len() as u32 * Self::field_size() as u32
-            + if include_size { Self::VEC_LEN_BYTES } else { 0 };
-
-        let mut res = Vec::with_capacity(total_size as usize);
-        if include_size {
-            Self::write_u32(&mut res, buf.len() as u32);
-        }
-        for el in buf.iter().cloned() {
-            Self::write_field_element(&mut res, el);
-        }
-        debug_assert_eq!(res.len(), total_size as usize);
-        res
-    }
-
-    pub fn read_u32(buf: &[u8], offset: &mut usize) -> u32 {
-        const BYTES: usize = 4;
-        let res = u32::from_be_bytes(buf[*offset..*offset + BYTES].try_into().unwrap());
-        *offset += BYTES;
-        res
-    }
-
-    pub(crate) fn read_u64(buf: &[u8], offset: &mut usize) -> u64 {
-        const BYTES: usize = 8;
-        let res = u64::from_be_bytes(buf[*offset..*offset + BYTES].try_into().unwrap());
-        *offset += BYTES;
-        res
-    }
-
-    pub fn read_biguint(buf: &[u8], num_64_limbs: usize, offset: &mut usize) -> BigUint {
-        let mut bigint = BigUint::default();
-        for _ in 0..num_64_limbs {
-            let data = Self::read_u64(buf, offset);
-            bigint <<= 64;
-            bigint += data;
-        }
-        bigint
-    }
-
-    pub fn write_u32(buf: &mut Vec<u8>, val: u32) {
-        buf.extend(val.to_be_bytes());
-    }
-
-    pub(crate) fn write_u64(buf: &mut Vec<u8>, val: u64) {
-        buf.extend(val.to_be_bytes());
-    }
-
-    pub fn write_field_element(buf: &mut Vec<u8>, el: F) {
-        let prev_len = buf.len();
-        for el in el.to_base_prime_field_elements() {
-            let el = el.into_bigint(); // Gets rid of montgomery form
-
-            for data in el.as_ref().iter().rev().cloned() {
-                Self::write_u64(buf, data);
-            }
-
-            debug_assert_eq!(
-                buf.len() - prev_len,
-                Self::FIELDSIZE_BYTES as usize * F::extension_degree() as usize
-            );
-        }
-    }
-
-    pub fn read_field_element(buf: &[u8], offset: &mut usize) -> F {
-        let mut fields = Vec::with_capacity(F::extension_degree() as usize);
-
-        for _ in 0..F::extension_degree() {
-            let mut bigint: BigUint = Default::default();
-            for _ in 0..Self::NUM_64_LIMBS {
-                let data = Self::read_u64(buf, offset);
-                bigint <<= 64;
-                bigint += data;
-            }
-            fields.push(F::BasePrimeField::from(bigint));
-        }
-
-        F::from_base_prime_field_elems(fields).expect("Should work")
-    }
 }
 
 impl<C: CurveGroup> SerializeC<C> {
@@ -161,11 +32,11 @@ impl<C: CurveGroup> SerializeC<C> {
         } else {
             let (x, y) = el.xy().unwrap_or_default();
             if write_x_first {
-                Serialize::write_field_element(buf, x);
-                Serialize::write_field_element(buf, y);
+                SerializeF::write_field_element(buf, x);
+                SerializeF::write_field_element(buf, y);
             } else {
-                Serialize::write_field_element(buf, y);
-                Serialize::write_field_element(buf, x);
+                SerializeF::write_field_element(buf, y);
+                SerializeF::write_field_element(buf, x);
             }
         }
 
@@ -189,8 +60,8 @@ impl<P: HonkCurve<TranscriptFieldType>> SerializeP<P> {
             *offset += Self::FIELDSIZE_BYTES as usize * 2;
             return P::Affine::zero();
         }
-        let first = Serialize::<P::BaseField>::read_field_element(buf, offset);
-        let second = Serialize::<P::BaseField>::read_field_element(buf, offset);
+        let first = SerializeF::<P::BaseField>::read_field_element(buf, offset);
+        let second = SerializeF::<P::BaseField>::read_field_element(buf, offset);
 
         if read_x_first {
             P::g1_affine_from_xy(first, second)
