+++ co-noir/common/src/lib.rs
@@ -3,10 +3,9 @@ use crate::{
     transcript::{Transcript, TranscriptFieldType, TranscriptHasher},
 };
 use ark_ec::CurveGroup;
-use ark_ff::PrimeField;
 use co_builder::{
     HonkProofResult,
-    prelude::{HonkCurve, ProverCrs, Serialize, Utils},
+    prelude::{HonkCurve, ProverCrs, Utils},
 };
 use mpc_net::Network;
 
@@ -98,40 +97,5 @@ impl CoUtils {
         state: &mut T::State,
     ) -> eyre::Result<()> {
         T::inv_many_in_place_leaking_zeros(poly, net, state)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct HonkProof<F: PrimeField> {
-    proof: Vec<F>,
-}
-
-impl<F: PrimeField> HonkProof<F> {
-    pub fn new(proof: Vec<F>) -> Self {
-        Self { proof }
-    }
-
-    pub fn inner(self) -> Vec<F> {
-        self.proof
-    }
-
-    pub fn to_buffer(&self) -> Vec<u8> {
-        Serialize::to_buffer(&self.proof, false)
-    }
-
-    pub fn from_buffer(buf: &[u8]) -> HonkProofResult<Self> {
-        let res = Serialize::from_buffer(buf, false)?;
-        Ok(Self::new(res))
-    }
-
-    pub fn separate_proof_and_public_inputs(self, num_public_inputs: usize) -> (Self, Vec<F>) {
-        let (public_inputs, proof) = self.proof.split_at(num_public_inputs);
-        (Self::new(proof.to_vec()), public_inputs.to_vec())
-    }
-
-    pub fn insert_public_inputs(self, public_inputs: Vec<F>) -> Self {
-        let mut proof = public_inputs;
-        proof.extend(self.proof.to_owned());
-        Self::new(proof)
     }
 }
