+++ co-noir/common/src/transcript.rs
@@ -1,11 +1,9 @@
-use crate::{
-    HonkProof,
-    sponge_hasher::{FieldHash, FieldSponge},
-};
+use crate::sponge_hasher::{FieldHash, FieldSponge};
 use ark_ec::AffineRepr;
 use ark_ff::{One, PrimeField, Zero};
 use co_builder::{HonkProofError, HonkProofResult, prelude::HonkCurve};
 use mpc_core::gadgets::poseidon2::Poseidon2;
+use noir_types::HonkProof;
 use num_bigint::BigUint;
 use serde::Deserialize;
 use serde::Serialize;
